+++ lib/src/model/top_level_variable.dart
@@ -80,4 +80,8 @@ class TopLevelVariable extends ModelElement
 
   @override
   Iterable<CommentReferable> get referenceParents => [definingLibrary];
+
+  @override
+  bool get hasHideConstantImplementation =>
+      definingLibrary.hasHideConstantImplementations;
 }